chore: add client tunne mod
