+++ vpn/src/service/client_tunnel/tunnel_read.rs
@@ -1,0 +1,28 @@
+use futures::channel::mpsc::{Receiver, Sender};
+
+use crate::{ClientToSocks5Msg, Socks5ToClientMsg};
+
+#[allow(unused)]
+pub struct TunnelReader {
+    pub id: u32,
+    pub tx: Sender<Socks5ToClientMsg>,
+    pub rx: Option<Receiver<ClientToSocks5Msg>>,
+}
+
+impl TunnelReader {
+    // pub async fn recv(&mut self) -> Option<ClientToSocks5Msg> {
+    //     self.rx.as_mut().unwrap().next().await
+    // }
+
+    // pub async fn send(&mut self, msg: Socks5ToClientMsg) -> Result<(), VpnError> {
+    //     self.tx
+    //         .send(msg)
+    //         .await
+    //         .expect("Send tunnel read port failed");
+    //     Ok(())
+    // }
+
+    pub fn get_id(&self) -> u32 {
+        self.id
+    }
+}
